Main: bugref:1909: Compilation issues
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@90926 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-all/QMTranslatorImpl.cpp
@@ -443,7 +443,7 @@ public:
                     && (!apszDisabm[i]  || !*apszDisabm[i]  || RTStrCmp(message.pszComment, apszDisabm[i]) == 0 ))
                 {
                     const std::vector<const char *> &vecTranslations = m_messageArray[iter->offset].vecTranslations;
-                    size_t idxPlural = plural(aNum);
+                    size_t idxPlural = (size_t)plural(aNum);
                     return vecTranslations[RT_MIN(idxPlural, vecTranslations.size() - 1)];
                 }
             }
